Changed the was Printer Models are loded
To load new printers Modify the Color_Printers.list and Monochrome_Printers.list to include your printer respectfully. The Script gets the Model of printer from the <title></title> Tag. Make sure that your model of print is in one of these two files and this program "Should" find the Toner